// packages/shared/src/types.d.ts
export type FormValidationRule = {
    type: string;
    value?: unknown;
    message?: string;
    soft?: boolean;
};